Add new props package.
